refactor(grpc): hoist the transcription family contract out of qwen3_asr
The TranscriptionFamily trait, the FAMILIES registry, and the two
validation-error types lived inside the qwen3_asr module, so the
router's generic transcription stages imported their model-agnostic
seam from a model-named path. Move the contract to a neutral
llm_multimodal::registry::transcription module; qwen3_asr now only
implements it. Nothing model-named remains in the router's imports,
and a second family lands as its own module plus one FAMILIES row.
